Extract MCP integration from core
