+++ packages/core/package.json
@@ -116,8 +116,6 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@modelcontextprotocol/sdk": "^1.30.0",
-    "undici": "^8.10.0",
     "yaml": "^2.9.0",
     "zod": "^4.4.3"
   },
