--- packages/core/src/mcp/abort.ts
@@ -1,16 +1,0 @@
-export function linkAbortSignal(
-  source: AbortSignal | undefined,
-  target: AbortController,
-): () => void {
-  if (source === undefined) {
-    return () => undefined;
-  }
-  if (source.aborted) {
-    target.abort(source.reason);
-    return () => undefined;
-  }
-
-  const onAbort = () => target.abort(source.reason);
-  source.addEventListener("abort", onAbort, { once: true });
-  return () => source.removeEventListener("abort", onAbort);
-}
