+++ packages/core/src/mcp/index.ts
@@ -1,4 +1,2 @@
-export * from "./client";
-export * from "./client-group";
 export { isMcpTool } from "./tool";
 export * from "./types";
