+++ packages/mcp/src/url-safety.ts
